+++ src/extension.ts
@@ -1,9 +1,7 @@
 import * as vscode from 'vscode';
-import { useAsTemplate } from './functions/use-as-template';
+import { cancelRequest, useAsTemplate } from './functions/use-as-template';
 
 /**
- * Extension activated
- *
  * @param ExtensionContract context
  * @return void
  */
@@ -12,7 +10,12 @@ export function activate(context: vscode.ExtensionContext) {
 
 	disposable = vscode.commands.registerCommand('gpt-template.useAsTemplate', async (fileUri) => useAsTemplate(fileUri));
 	context.subscriptions.push(disposable);
+
+	disposable = vscode.commands.registerCommand('gpt-template.cancelRequest', async () => cancelRequest());
+	context.subscriptions.push(disposable);
 }
 
-// this method is called when your extension is deactivated
+/**
+ * @return void
+ */
 export function deactivate() {}