+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [0.3.0]
+
+-   Ability to cancel active generations (in status bar)
+-   Open new file if nothing is open
+-   Write to existing file if it's empty
+-   Optionally save to relative path
+
+## [0.2.0]
+
+-   Update app icon
+-   Add GPT streaming
+-   Update example GIF
+-   Use Settings class
+
 ## [Unreleased]
 
 -   Initial release